+++ CHANGELOG.md
@@ -1,9 +1,14 @@
+
 - Added: Initial support for Service Worker
   ([#343](https://github.com/MoOx/phenomic/pull/343))
 - Changed: Remove the ability to define a custom globby pattern via appcache option.
   `appcache` option was deprecated and will be removed soon.
   You should update your configuration to use the latest `offline` option.
   ([#343](https://github.com/MoOx/phenomic/pull/343))
+
+## Boilerplate
+
+- Removed: ``layouts/index.js``. It is not used since 0.10.0.
 
 # 0.10.2 - 2016-04-14