+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-
+- Changed: Normalize head date (server: Date, client: string)
+  ([#397](https://github.com/MoOx/phenomic/issues/397))
 - Added: Initial support for Service Worker
   ([#343](https://github.com/MoOx/phenomic/pull/343))
 - Changed: Remove the ability to define a custom globby pattern via appcache option.